+++ R/statistic-wmean.R
@@ -1,15 +1,24 @@
-#' Weighted mean
+#' Weighted Mean (WMEAN)
 #'
-#' Calculate the activity of all regulons in `network` through the conditions in
-#' the `mat` matrix by calculating the mean over the expression of all genes.
+#' @description
+#' Calculates regulatory activities by computing the WMEAN.
 #'
 #' @details
+<<<<<<< HEAD:R/statistic-mean.R
 #'  `run_mean()` calculates the activity score, but in addition, it takes
 #'  advantage of the permutations used to calculate the `p-value`, to provide
 #'  the normalized activity score and a corrected mean score.
 #'  This is represented in the `statistic` column which will contain three
 #'  values for each call to `run_mean()`; __mean__, __normalized_mean__, and
 #'  __corrected_mean__.
+=======
+#' Infers activity score for each regulator by weighting the molecular readouts
+#' of its targets by their mode of regulations and likelihoods. In addition, it
+#' runs permutations to calculate empirical p-values, providing normalized
+#' (z-score) and corrected activity (estimate * -log10(pval)) scores. This is
+#' represented in the `statistic` column which will contain three values for
+#' each call to `run_wmean()`; __wmean__, __norm_wmean__ and __corr_wmean__.
+>>>>>>> 079063fd42ac39301d8629ecfbdc606f5fe27cee:R/statistic-wmean.R
 #'
 #' @inheritParams .decoupler_mat_format
 #' @inheritParams .decoupler_network_format
@@ -19,13 +28,13 @@
 #' @param sparse Should the matrices used for the calculation be sparse?
 #' @param randomize_type How to randomize the expression matrix.
 #'
-#' @return A long format tibble of the enrichment scores for each tf
+#' @return A long format tibble of the enrichment scores for each source
 #'  across the samples. Resulting tibble contains the following columns:
 #'  1. `statistic`: Indicates which method is associated with which score.
-#'  2. `tf`: Source nodes of `network`.
+#'  2. `source`: Source nodes of `network`.
 #'  3. `condition`: Condition representing each column of `mat`.
 #'  4. `score`: Regulatory activity (enrichment score).
-#'  5. `p_value`: p-value for the score of mean method.
+#'  5. `p_value`: p-value for the score of the method.
 #' @family decoupleR statistics
 #' @export
 #' @import dplyr
@@ -39,14 +48,14 @@
 #' mat <- readRDS(file.path(inputs_dir, "input-expr_matrix.rds"))
 #' network <- readRDS(file.path(inputs_dir, "input-dorothea_genesets.rds"))
 #'
-#' run_mean(mat, network, tf, target, mor, likelihood)
-run_mean <- function(mat,
+#' run_wmean(mat, network, .source='tf')
+run_wmean <- function(mat,
                      network,
-                     .source = .data$tf,
+                     .source = .data$source,
                      .target = .data$target,
                      .mor = .data$mor,
                      .likelihood = .data$likelihood,
-                     times = 2,
+                     times = 100,
                      seed = 42,
                      sparse = TRUE,
                      randomize_type = "rows") {
@@ -55,19 +64,20 @@ run_mean <- function(mat,
         rlang::abort(message = stringr::str_glue("Parameter 'times' must be greater than or equal to 2, but {times} was passed."))
     }
 
+    # Check for NAs/Infs in mat
+    check_nas_infs(mat)
+
     network <- network %>%
-        convert_to_mean({{ .source }}, {{ .target }}, {{ .mor }}, {{ .likelihood }})
+        convert_to_wmean({{ .source }}, {{ .target }}, {{ .mor }}, {{ .likelihood }})
 
     # Preprocessing -----------------------------------------------------------
 
     # Calculate the weights that will be used for the evaluation of the model
     network <- network %>%
         filter(.data$target %in% rownames(mat)) %>%
-        .mean_calculate_weight()
+        .wmean_calculate_weight()
 
     # Extract labels that will map to the expression and profile matrices
-    tfs <- unique(network[["tf"]])
-
     shared_targets <- unique(network[["target"]])
 
     targets <- rownames(mat)
@@ -76,7 +86,7 @@ run_mean <- function(mat,
     # Extract matrix of weights
     weight_mat <- network %>%
         pivot_wider_profile(
-            id_cols = .data$tf,
+            id_cols = .data$source,
             names_from = .data$target,
             values_from = .data$weight,
             to_matrix = TRUE,
@@ -84,34 +94,39 @@ run_mean <- function(mat,
             values_fill = 0
         )
 
+    weight_mat <- as.matrix(weight_mat)
+
+    # This fixes the wrong denominator defined in contribution
+    weight_mat <- weight_mat/rowSums(abs(weight_mat))
+
     # Analysis ----------------------------------------------------------------
     withr::with_seed(seed, {
-        .mean_analysis(mat, weight_mat, shared_targets, times, randomize_type)
+        .wmean_analysis(mat, weight_mat, shared_targets, times, randomize_type)
     })
 
 }
 
 # Helper functions --------------------------------------------------------
 
-#' Wrapper to execute run_mean() logic once finished preprocessing of data
+#' Wrapper to execute run_wmean() logic once finished preprocessing of data
 #'
-#' @inherit run_mean description
+#' @inherit run_wmean description
 #'
-#' @inheritParams run_mean
+#' @inheritParams run_wmean
 #' @param weight_mat Matrix that corresponds to the multiplication of the mor
 #'  column with likelihood divided over the contribution.
 #' @param shared_targets Target nodes that are shared between the
 #'  `mat` and `network`.
 #'
-#' @inherit run_mean return
+#' @inherit run_wmean return
 #'
 #' @keywords internal
 #' @noRd
-.mean_analysis <- function(mat, weight_mat, shared_targets, times, randomize_type) {
+.wmean_analysis <- function(mat, weight_mat, shared_targets, times, randomize_type) {
     # Thus, it is only necessary to define if we want
     # to evaluate a random model or not.
-    mean_run <- partial(
-        .mean_run,
+    wmean_run <- partial(
+        .wmean_run,
         mat = mat,
         weight_mat = weight_mat,
         shared_targets = shared_targets,
@@ -119,8 +134,8 @@ run_mean <- function(mat,
     )
 
     # Run model for random data
-    map_dfr(seq_len(times), ~ mean_run(random = TRUE)) %>%
-        group_by(.data$tf, .data$condition) %>%
+    map_dfr(seq_len(times), ~ wmean_run(random = TRUE)) %>%
+        group_by(.data$source, .data$condition) %>%
         summarise(
             null_distribution = list(.data$value),
             null_mean = mean(.data$value),
@@ -128,7 +143,7 @@ run_mean <- function(mat,
             .groups = "drop"
         ) %>%
         # Run the true model and joined to random.
-        left_join(y = mean_run(random = FALSE), by = c("tf", "condition")) %>%
+        left_join(y = wmean_run(random = FALSE), by = c("source", "condition")) %>%
         # Calculate scores
         mutate(
             z_score = (.data$value - .data$null_mean) / .data$null_sd,
@@ -138,44 +153,53 @@ run_mean <- function(mat,
                 .y = .data$value,
                 .f = ~ sum(abs(.x) > abs(.y)) / length(.x)
             ),
+<<<<<<< HEAD:R/statistic-mean.R
             p_value = if_else(.data$p_value == 0, 1 / times, .data$p_value),
             corrected_mean = - .data$value * log10(.data$p_value)
+=======
+            c_p_value = ifelse(.data$p_value == 0, 1/length(.data$null_distribution), .data$p_value),
+            c_score = .data$value * (-log10(.data$c_p_value))
+>>>>>>> 079063fd42ac39301d8629ecfbdc606f5fe27cee:R/statistic-wmean.R
         ) %>%
         # Reformat results
         select(-contains("null")) %>%
-        rename(mean = .data$value, normalized_mean = .data$z_score) %>%
+        rename(corr_wmean = .data$c_score, wmean = .data$value, norm_wmean = .data$z_score) %>%
         pivot_longer(
+<<<<<<< HEAD:R/statistic-mean.R
             cols = c(.data$mean, .data$normalized_mean, .data$corrected_mean),
+=======
+            cols = c(.data$corr_wmean, .data$wmean, .data$norm_wmean),
+>>>>>>> 079063fd42ac39301d8629ecfbdc606f5fe27cee:R/statistic-wmean.R
             names_to = "statistic",
             values_to = "score"
         ) %>%
-        arrange(.data$statistic, .data$tf, .data$condition) %>%
-        select(.data$statistic, .data$tf, .data$condition, .data$score, .data$p_value)
+        arrange(.data$statistic, .data$source, .data$condition) %>%
+        select(.data$statistic, .data$source, .data$condition, .data$score, .data$p_value)
 }
 
-#' Wrapper to run mean one time
+#' Wrapper to run wmean one time
 #'
-#' @inheritParams .mean_analysis
-#' @inherit .mean_evaluate_model return
+#' @inheritParams .wmean_analysis
+#' @inherit .wmean_evaluate_model return
 #' @keywords internal
 #' @noRd
-.mean_run <- function(mat, weight_mat, shared_targets, random, randomize_type) {
-    .mean_map_model_data(mat, shared_targets, random, randomize_type) %>%
-        .mean_evaluate_model(weight_mat)
+.wmean_run <- function(mat, weight_mat, shared_targets, random, randomize_type) {
+    .wmean_map_model_data(mat, shared_targets, random, randomize_type) %>%
+        .wmean_evaluate_model(weight_mat)
 }
 
 #' Calculate mean weight
 #'
-#' @inheritParams .mean_analysis
+#' @inheritParams .wmean_analysis
 #' @keywords internal
 #' @noRd
-.mean_calculate_weight <- function(network) {
+.wmean_calculate_weight <- function(network) {
     network %>%
-        add_count(.data$tf, name = "contribution") %>%
+        add_count(.data$source, name = "contribution") %>%
         transmute(
-            .data$tf,
+            .data$source,
             .data$target,
-            weight = .data$mor * .data$likelihood / .data$contribution
+            weight = .data$mor * .data$likelihood
         )
 }
 
@@ -188,10 +212,10 @@ run_mean <- function(mat,
 #'
 #' @return Matrix with rows that match `shared_targets`.
 #'
-#' @inheritParams .mean_analysis
+#' @inheritParams .wmean_analysis
 #' @keywords internal
 #' @noRd
-.mean_map_model_data <- function(mat, shared_targets, random, randomize_type) {
+.wmean_map_model_data <- function(mat, shared_targets, random, randomize_type) {
     if (random) {
         randomize_matrix(mat, randomize_type = randomize_type)[shared_targets, ]
     } else {
@@ -205,17 +229,17 @@ run_mean <- function(mat,
 #' weights by the factor of interest and comparing it against results
 #' from permutations of the matrix of values of interest.
 #'
-#' @inheritParams .mean_analysis
+#' @inheritParams .wmean_analysis
 #'
 #' @return A dataframe with three columns:
-#'  tf (source nodes), condition (colnames of mat) and value (score).
+#'  source (source nodes), condition (colnames of mat) and value (score).
 #'
 #' @keywords internal
 #' @noRd
-.mean_evaluate_model <- function(mat, weight_mat) {
+.wmean_evaluate_model <- function(mat, weight_mat) {
     (weight_mat %*% mat) %>%
         as.matrix() %>%
         as.data.frame() %>%
-        rownames_to_column("tf") %>%
-        pivot_longer(-.data$tf, names_to = "condition")
+        rownames_to_column("source") %>%
+        pivot_longer(-.data$source, names_to = "condition")
 }